@@ -3,16 +3,13 @@ using Contracts;
 using Entities.DataTransferObjects;
 using Entities.Models;
 using Microsoft.AspNetCore.Mvc;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 
 namespace AccountOwnerServer.Controllers
 {
-    [Route("api/owner")] 
-    [ApiController] 
-    public class OwnerController : ControllerBase 
-    { 
+    [Route("api/owner")]
+    [ApiController]
+    public class OwnerController : ControllerBase
+    {
         private ILoggerManager _logger; 
         private IRepositoryWrapper _repository;
         private IMapper _mapper;
@@ -20,10 +17,10 @@ namespace AccountOwnerServer.Controllers
         public OwnerController(ILoggerManager logger, IRepositoryWrapper repository, IMapper mapper) 
         { 
             _logger = logger; 
-            _repository = repository;
+            _repository = repository; 
             _mapper = mapper;
         }
-        
+
         [HttpGet] 
         public IActionResult GetAllOwners() 
         { 
@@ -38,68 +35,69 @@ namespace AccountOwnerServer.Controllers
             catch (Exception ex) 
             { 
                 _logger.LogError($"Something went wrong inside GetAllOwners action: {ex.Message}"); 
+                
                 return StatusCode(500, "Internal server error"); 
             } 
         }
 
-        [HttpGet("{id}", Name = "OwnerById")] 
-        public IActionResult GetOwnerById(Guid id) 
-        { 
-            try 
-            { 
-                var owner = _repository.Owner.GetOwnerById(id); 
-                if (owner == null) 
-                { 
-                    _logger.LogError($"Owner with id: {id}, hasn't been found in db."); 
-                    return NotFound(); 
-                } 
-                else 
-                { 
+        [HttpGet("{id}", Name = "OwnerById")]
+        public IActionResult GetOwnerById(Guid id)
+        {
+            try
+            {
+                var owner = _repository.Owner.GetOwnerById(id);
+                if (owner is null)
+                {
+                    _logger.LogError($"Owner with id: {id}, hasn't been found in db.");
+                    return NotFound();
+                }
+                else
+                {
                     _logger.LogInfo($"Returned owner with id: {id}");
 
                     var ownerResult = _mapper.Map<OwnerDto>(owner);
-                    return Ok(ownerResult); 
-                } 
-            } 
-            catch (Exception ex) 
-            { 
-                _logger.LogError($"Something went wrong inside GetOwnerById action: {ex.Message}"); 
-                return StatusCode(500, "Internal server error"); 
-            } 
+                    return Ok(ownerResult);
+                }
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError($"Something went wrong inside GetOwnerById action: {ex.Message}");
+                return StatusCode(500, "Internal server error");
+            }
         }
 
-        [HttpGet("{id}/account")] 
-        public IActionResult GetOwnerWithDetails(Guid id) 
-        { 
-            try 
-            { 
-                var owner = _repository.Owner.GetOwnerWithDetails(id); 
-                if (owner == null) 
-                { 
-                    _logger.LogError($"Owner with id: {id}, hasn't been found in db."); 
-                    return NotFound(); 
-                } 
-                else 
-                { 
+        [HttpGet("{id}/account")]
+        public IActionResult GetOwnerWithDetails(Guid id)
+        {
+            try
+            {
+                var owner = _repository.Owner.GetOwnerWithDetails(id);
+                if (owner == null)
+                {
+                    _logger.LogError($"Owner with id: {id}, hasn't been found in db.");
+                    return NotFound();
+                }
+                else
+                {
                     _logger.LogInfo($"Returned owner with details for id: {id}");
 
                     var ownerResult = _mapper.Map<OwnerDto>(owner);
-                    return Ok(ownerResult); 
-                } 
-            } 
-            catch (Exception ex) 
-            { 
-                _logger.LogError($"Something went wrong inside GetOwnerWithDetails action: {ex.Message}"); 
-                return StatusCode(500, "Internal server error"); 
+                    return Ok(ownerResult);
+                }
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError($"Something went wrong inside GetOwnerWithDetails action: {ex.Message}");
+                return StatusCode(500, "Internal server error");
             }
         }
 
         [HttpPost]
-        public IActionResult CreateOwner([FromBody]OwnerForCreationDto owner)
+        public IActionResult CreateOwner([FromBody] OwnerForCreationDto owner)
         {
             try
             {
-                if (owner == null)
+                if (owner is null)
                 {
                     _logger.LogError("Owner object sent from client is null.");
                     return BadRequest("Owner object is null");
@@ -128,11 +126,11 @@ namespace AccountOwnerServer.Controllers
         }
 
         [HttpPut("{id}")]
-        public IActionResult UpdateOwner(Guid id, [FromBody]OwnerForUpdateDto owner)
+        public IActionResult UpdateOwner(Guid id, [FromBody] OwnerForUpdateDto owner)
         {
             try
             {
-                if (owner == null)
+                if (owner is null)
                 {
                     _logger.LogError("Owner object sent from client is null.");
                     return BadRequest("Owner object is null");
@@ -145,7 +143,7 @@ namespace AccountOwnerServer.Controllers
                 }
 
                 var ownerEntity = _repository.Owner.GetOwnerById(id);
-                if (ownerEntity == null)
+                if (ownerEntity is null)
                 {
                     _logger.LogError($"Owner with id: {id}, hasn't been found in db.");
                     return NotFound();
@@ -177,10 +175,10 @@ namespace AccountOwnerServer.Controllers
                     return NotFound();
                 }
 
-                if (_repository.Account.AccountsByOwner(id).Any()) 
+                if (_repository.Account.AccountsByOwner(id).Any())
                 {
-                    _logger.LogError($"Cannot delete owner with id: {id}. It has related accounts. Delete those accounts first"); 
-                    return BadRequest("Cannot delete owner. It has related accounts. Delete those accounts first"); 
+                    _logger.LogError($"Cannot delete owner with id: {id}. It has related accounts. Delete those accounts first");
+                    return BadRequest("Cannot delete owner. It has related accounts. Delete those accounts first");
                 }
 
                 _repository.Owner.DeleteOwner(owner);
