@@ -1,10 +1,7 @@
 ﻿using Contracts;
-using Entities;
 using LoggerService;
-using Microsoft.AspNetCore.Builder;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
+using Entities;
 using Repository;
 
 namespace AccountOwnerServer.Extensions
@@ -30,15 +27,17 @@ namespace AccountOwnerServer.Extensions
             });
         }
 
-        public static void ConfigureLoggerService(this IServiceCollection services)
+        public static void ConfigureLoggerService(this IServiceCollection services) 
         {
-            services.AddSingleton<ILoggerManager, LoggerManager>();
+            services.AddSingleton<ILoggerManager, LoggerManager>(); 
         }
 
-        public static void ConfigureMySqlContext(this IServiceCollection services, IConfiguration config) 
-        { 
-            var connectionString = config["mysqlconnection:connectionString"]; 
-            services.AddDbContext<RepositoryContext>(o => o.UseMySql(connectionString)); 
+        public static void ConfigureMySqlContext(this IServiceCollection services, IConfiguration config)
+        {
+            var connectionString = config["mysqlconnection:connectionString"];
+
+            services.AddDbContext<RepositoryContext>(o => o.UseMySql(connectionString, 
+                MySqlServerVersion.LatestSupportedServerVersion));
         }
 
         public static void ConfigureRepositoryWrapper(this IServiceCollection services) 
