@@ -6,6 +6,6 @@ namespace Entities.DataTransferObjects
     {
         public Guid Id { get; set; }
         public DateTime DateCreated { get; set; }
-        public string AccountType { get; set; }
+        public string? AccountType { get; set; }
     }
 }
