@@ -1,17 +1,12 @@
-﻿using Entities.Models;
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Entities.DataTransferObjects
+﻿namespace Entities.DataTransferObjects
 {
     public class OwnerDto
     {
         public Guid Id { get; set; }
-        public string Name { get; set; }
+        public string? Name { get; set; }
         public DateTime DateOfBirth { get; set; }
-        public string Address { get; set; }
+        public string? Address { get; set; }
 
-        public IEnumerable<AccountDto> Accounts { get; set; }
+        public IEnumerable<AccountDto>? Accounts { get; set; }
     }
 }
