@@ -10,7 +10,7 @@ namespace Entities
         { 
         }
 
-        public DbSet<Owner> Owners { get; set; }
-        public DbSet<Account> Accounts { get; set; }
+        public DbSet<Owner>? Owners { get; set; }
+        public DbSet<Account>? Accounts { get; set; }
     }
 }
