@@ -1,8 +1,6 @@
 ﻿using Contracts;
 using Entities;
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Linq;
 using System.Linq.Expressions;
 
 namespace Repository
@@ -13,31 +11,17 @@ namespace Repository
         public RepositoryBase(RepositoryContext repositoryContext) 
         {
             RepositoryContext = repositoryContext; 
-        } 
-        
-        public IQueryable<T> FindAll() 
-        { 
-            return RepositoryContext.Set<T>().AsNoTracking(); 
-        } 
-        
-        public IQueryable<T> FindByCondition(Expression<Func<T, bool>> expression)
-        { 
-            return RepositoryContext.Set<T>().Where(expression).AsNoTracking();
-        } 
-        
-        public void Create(T entity) 
-        { 
-            RepositoryContext.Set<T>().Add(entity); 
-        } 
-        
-        public void Update(T entity) 
-        {
-            RepositoryContext.Set<T>().Update(entity);
-        } 
-        
-        public void Delete(T entity)
-        { 
-            RepositoryContext.Set<T>().Remove(entity); 
         }
+
+        public IQueryable<T> FindAll() => RepositoryContext.Set<T>().AsNoTracking();
+
+        public IQueryable<T> FindByCondition(Expression<Func<T, bool>> expression) => 
+            RepositoryContext.Set<T>().Where(expression).AsNoTracking();
+
+        public void Create(T entity) => RepositoryContext.Set<T>().Add(entity);
+
+        public void Update(T entity) => RepositoryContext.Set<T>().Update(entity);
+
+        public void Delete(T entity) => RepositoryContext.Set<T>().Remove(entity);
     }
 }
