@@ -1,6 +1,6 @@
 export interface Account{
-    id: string;
-    dateCreated: Date;
-    accountType: string;
-    ownerId?: string;
+  id: string;
+  dateCreated: Date;
+  accountType: string;
+  ownerId?: string;
 }
