@@ -1,9 +1,9 @@
 import { Account } from './account.model';
 export interface Owner{
-    id: string;
-    name: string;
-    dateOfBirth: Date;
-    address: string;
+  id: string;
+  name: string;
+  dateOfBirth: Date;
+  address: string;
 
-    accounts?: Account[];
+  accounts?: Account[];
 }
