@@ -6,7 +6,7 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./internal-server.component.css']
 })
 export class InternalServerComponent implements OnInit {
-  public errorMessage: string = "500 SERVER ERROR, CONTACT ADMINISTRATOR!!!!";
+  errorMessage: string = "500 SERVER ERROR, CONTACT ADMINISTRATOR!!!!";
   
   constructor() { }
 
