@@ -6,7 +6,7 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./not-found.component.css']
 })
 export class NotFoundComponent implements OnInit {
-  public notFoundText: string = `404 SORRY COULDN'T FIND IT!!!`
+  notFoundText: string = `404 SORRY COULDN'T FIND IT!!!`;
 
   constructor() { }
 
