@@ -6,12 +6,12 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./home.component.css']
 })
 export class HomeComponent implements OnInit {
+
   public homeText: string;
 
   constructor() { }
 
   ngOnInit(): void {
-    this.homeText = "WELCOME TO ACCOUNT-OWNER APPLICATION"
+    this.homeText = "WELCOME TO ACCOUNT-OWNER APPLICATION";
   }
-
 }
