@@ -1,76 +1,86 @@
+import { SuccessModalComponent } from './../../shared/modals/success-modal/success-modal.component';
+import { DatePipe } from '@angular/common';
+import { Router } from '@angular/router';
+import { ErrorHandlerService } from './../../shared/services/error-handler.service';
+import { OwnerRepositoryService } from './../../shared/services/owner-repository.service';
 import { Component, OnInit } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
-import { OwnerForCreation } from './../../_interfaces/ownerForCreation.model';
-import { ErrorHandlerService } from './../../shared/services/error-handler.service';
-import { RepositoryService } from './../../shared/services/repository.service';
-import { Router } from '@angular/router';
-import { DatePipe } from '@angular/common';
- 
+import { Owner } from 'src/app/_interfaces/owner.model';
+import { HttpErrorResponse } from '@angular/common/http';
+import { OwnerForCreation } from 'src/app/_interfaces/ownerForCreation.model';
+import { ModalOptions, BsModalRef, BsModalService } from 'ngx-bootstrap/modal';
+
 @Component({
   selector: 'app-owner-create',
   templateUrl: './owner-create.component.html',
   styleUrls: ['./owner-create.component.css']
 })
 export class OwnerCreateComponent implements OnInit {
-  public errorMessage: string = '';
- 
-  public ownerForm: FormGroup;
- 
-  constructor(private repository: RepositoryService, private errorHandler: ErrorHandlerService, private router: Router, private datePipe: DatePipe) { }
- 
-  ngOnInit() {
+  errorMessage: string = '';
+  ownerForm: FormGroup;
+  bsModalRef?: BsModalRef;
+
+  constructor(private repository: OwnerRepositoryService, private errorHandler: ErrorHandlerService,
+    private router: Router, private datePipe: DatePipe, private modal: BsModalService) { }
+
+  ngOnInit(): void {
     this.ownerForm = new FormGroup({
       name: new FormControl('', [Validators.required, Validators.maxLength(60)]),
       dateOfBirth: new FormControl('', [Validators.required]),
       address: new FormControl('', [Validators.required, Validators.maxLength(100)])
     });
   }
- 
-  public validateControl = (controlName: string) => {
-    if (this.ownerForm.controls[controlName].invalid && this.ownerForm.controls[controlName].touched)
+
+  validateControl = (controlName: string) => {
+    if (this.ownerForm.get(controlName).invalid && this.ownerForm.get(controlName).touched)
       return true;
- 
+    
+    return false;
+  } 
+
+  hasError = (controlName: string, errorName: string) => {
+    if (this.ownerForm.get(controlName).hasError(errorName))
+      return true;
+    
     return false;
   }
- 
-  public hasError = (controlName: string, errorName: string) => {
-    if (this.ownerForm.controls[controlName].hasError(errorName))
-      return true;
- 
-    return false;
-  }
- 
-  public executeDatePicker = (event) => {
-    this.ownerForm.patchValue({ 'dateOfBirth': event });
-  }
- 
-  public createOwner = (ownerFormValue) => {
-    if (this.ownerForm.valid) {
+
+  createOwner = (ownerFormValue) => {
+    if (this.ownerForm.valid)
       this.executeOwnerCreation(ownerFormValue);
-    }
   }
- 
+
   private executeOwnerCreation = (ownerFormValue) => {
     const owner: OwnerForCreation = {
       name: ownerFormValue.name,
       dateOfBirth: this.datePipe.transform(ownerFormValue.dateOfBirth, 'yyyy-MM-dd'),
       address: ownerFormValue.address
     }
- 
     const apiUrl = 'api/owner';
-    this.repository.create(apiUrl, owner)
-      .subscribe(res => {
-        $('#successModal').modal();
+    this.repository.createOwner(apiUrl, owner)
+    .subscribe({
+      next: (own: Owner) => {
+        const config: ModalOptions = {
+          initialState: {
+            modalHeaderText: 'Success Message',
+            modalBodyText: `Owner: ${own.name} created successfully`,
+            okButtonText: 'OK'
+          }
+        };
+
+        this.bsModalRef = this.modal.show(SuccessModalComponent, config);
+        this.bsModalRef.content.redirectOnOk.subscribe(_ => this.redirectToOwnerList());
       },
-      (error => {
-        this.errorHandler.handleError(error);
-        this.errorMessage = this.errorHandler.errorMessage;
-      })
-    )
+      error: (err: HttpErrorResponse) => {
+          this.errorHandler.handleError(err);
+          this.errorMessage = this.errorHandler.errorMessage;
+      }
+    })
+
   }
- 
-  public redirectToOwnerList(){
-    this.router.navigate(['/owner/list']);
-  }
- 
+
+redirectToOwnerList = () => {
+  this.router.navigate(['/owner/list']);
+}
+
 }
