@@ -1,53 +1,62 @@
-import { Component, OnInit } from '@angular/core';
-import { ErrorHandlerService } from './../../shared/services/error-handler.service';
-import { RepositoryService } from './../../shared/services/repository.service';
-import { Owner } from './../../_interfaces/owner.model';
+import { HttpErrorResponse } from '@angular/common/http';
 import { Router, ActivatedRoute } from '@angular/router';
- 
+import { ErrorHandlerService } from './../../shared/services/error-handler.service';
+import { OwnerRepositoryService } from './../../shared/services/owner-repository.service';
+import { Owner } from './../../_interfaces/owner.model';
+import { Component, OnInit } from '@angular/core';
+import { BsModalRef, ModalOptions, BsModalService } from 'ngx-bootstrap/modal';
+import { SuccessModalComponent } from 'src/app/shared/modals/success-modal/success-modal.component';
+
 @Component({
   selector: 'app-owner-delete',
   templateUrl: './owner-delete.component.html',
   styleUrls: ['./owner-delete.component.css']
 })
 export class OwnerDeleteComponent implements OnInit {
-  public errorMessage: string = '';
-  public owner: Owner;
- 
-constructor(private repository: RepositoryService, private errorHandler: ErrorHandlerService, private router: Router,
-  private activeRoute: ActivatedRoute) { }
- 
-  ngOnInit() {
+  owner: Owner;
+  bsModalRef?: BsModalRef;
+
+  constructor(private repository: OwnerRepositoryService, private errorHandler: ErrorHandlerService,
+    private router: Router, private activeRoute: ActivatedRoute, private modal: BsModalService) { }
+
+  ngOnInit(): void {
     this.getOwnerById();
   }
-   
+
   private getOwnerById = () => {
     const ownerId: string = this.activeRoute.snapshot.params['id'];
-    const ownerByIdUrl: string = `api/owner/${ownerId}`;
-   
-    this.repository.getData(ownerByIdUrl)
-      .subscribe(res => {
-        this.owner = res as Owner;
-      },
-      (error) => {
-        this.errorHandler.handleError(error);
-        this.errorMessage = this.errorHandler.errorMessage;
-      })
+    const apiUri: string = `api/owner/${ownerId}`;
+
+    this.repository.getOwner(apiUri)
+    .subscribe({
+      next: (own: Owner) => this.owner = own,
+      error: (err: HttpErrorResponse) => this.errorHandler.handleError(err)
+    })
   }
-   
-  public redirectToOwnerList = () => {
+
+  deleteOwner = () => {
+    const deleteUri: string = `api/owner/${this.owner.id}`;
+
+    this.repository.deleteOwner(deleteUri)
+    .subscribe({
+      next: (_) => {
+        const config: ModalOptions = {
+          initialState: {
+            modalHeaderText: 'Success Message',
+            modalBodyText: `Owner deleted successfully`,
+            okButtonText: 'OK'
+          }
+        };
+
+        this.bsModalRef = this.modal.show(SuccessModalComponent, config);
+        this.bsModalRef.content.redirectOnOk.subscribe(_ => this.redirectToOwnerList());
+      },
+      error: (err: HttpErrorResponse) => this.errorHandler.handleError(err)
+    })
+  }
+
+  redirectToOwnerList = () => {
     this.router.navigate(['/owner/list']);
   }
 
-  public deleteOwner = () => {
-    const deleteUrl: string = `api/owner/${this.owner.id}`;
-    this.repository.delete(deleteUrl)
-      .subscribe(res => {
-        $('#successModal').modal();
-      },
-      (error) => {
-        this.errorHandler.handleError(error);
-        this.errorMessage = this.errorHandler.errorMessage;
-      })
-  }
 }
-
