@@ -1,34 +1,33 @@
-import { SharedModule } from './../shared/shared.module';
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { RouterModule } from '@angular/router';
+import { OwnerRoutingModule } from './owner-routing.module';
 import { ReactiveFormsModule } from '@angular/forms';
- 
+import { BsDatepickerModule } from 'ngx-bootstrap/datepicker';
+
 import { OwnerListComponent } from './owner-list/owner-list.component';
 import { OwnerDetailsComponent } from './owner-details/owner-details.component';
+import { SharedModule } from '../shared/shared.module';
+import { OwnerAccountsComponent } from './owner-details/owner-accounts/owner-accounts.component';
 import { OwnerCreateComponent } from './owner-create/owner-create.component';
 import { OwnerUpdateComponent } from './owner-update/owner-update.component';
 import { OwnerDeleteComponent } from './owner-delete/owner-delete.component';
- 
+
+
 @NgModule({
-  imports: [
-    CommonModule,
-    SharedModule,
-    ReactiveFormsModule,
-    RouterModule.forChild([
-      { path: 'list', component: OwnerListComponent },
-      { path: 'details/:id', component: OwnerDetailsComponent },
-      { path: 'create', component: OwnerCreateComponent },
-      { path: 'update/:id', component: OwnerUpdateComponent },
-      { path: 'delete/:id', component: OwnerDeleteComponent }
-    ])
-  ],
   declarations: [
     OwnerListComponent,
     OwnerDetailsComponent,
+    OwnerAccountsComponent,
     OwnerCreateComponent,
     OwnerUpdateComponent,
     OwnerDeleteComponent
+  ],
+  imports: [
+    CommonModule,
+    OwnerRoutingModule,
+    SharedModule,
+    ReactiveFormsModule,
+    BsDatepickerModule.forRoot()
   ]
 })
 export class OwnerModule { }
