@@ -1,16 +1,17 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
+import { BsModalRef } from 'ngx-bootstrap/modal';
 
 @Component({
   selector: 'app-error-modal',
   templateUrl: './error-modal.component.html',
-  styleUrls: ['./error-modal.component.css', '../modal-shared.component.css']
+  styleUrls: ['./error-modal.component.css']
 })
 export class ErrorModalComponent implements OnInit {
-  @Input() public modalHeaderText: string;
-  @Input() public modalBodyText: string;
-  @Input() public okButtonText: string;
+  modalHeaderText: string;
+  modalBodyText: string;
+  okButtonText: string;
 
-  constructor() { }
+  constructor(public bsModalRef: BsModalRef) { }
 
   ngOnInit(): void {
   }
