@@ -1,5 +1,7 @@
-import { Injectable } from '@angular/core';
+import { ErrorModalComponent } from './../modals/error-modal/error-modal.component';
+import { BsModalService, ModalOptions } from 'ngx-bootstrap/modal';
 import { HttpErrorResponse } from '@angular/common/http';
+import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
 
 @Injectable({
@@ -7,36 +9,43 @@ import { Router } from '@angular/router';
 })
 export class ErrorHandlerService {
   public errorMessage: string = '';
- 
-  constructor(private router: Router) { }
- 
+
+  constructor(private router: Router, private modal: BsModalService) { }
+
   public handleError = (error: HttpErrorResponse) => {
-    if(error.status === 500){
+    if (error.status === 500) {
       this.handle500Error(error);
     }
-    else if(error.status === 404){
+    else if (error.status === 404) {
       this.handle404Error(error)
     }
-    else{
+    else {
       this.handleOtherError(error);
     }
   }
- 
+
   private handle500Error = (error: HttpErrorResponse) => {
     this.createErrorMessage(error);
     this.router.navigate(['/500']);
   }
- 
+
   private handle404Error = (error: HttpErrorResponse) => {
     this.createErrorMessage(error);
     this.router.navigate(['/404']);
   }
- 
   private handleOtherError = (error: HttpErrorResponse) => {
     this.createErrorMessage(error);
-    $('#errorModal').modal();
+
+    const config: ModalOptions = {
+      initialState: {
+        modalHeaderText: 'Error Message',
+        modalBodyText: this.errorMessage,
+        okButtonText: 'OK'
+      }
+    };
+    this.modal.show(ErrorModalComponent, config);
   }
- 
+
   private createErrorMessage = (error: HttpErrorResponse) => {
     this.errorMessage = error.error ? error.error : error.statusText;
   }
