@@ -1,20 +1,25 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { ModalModule } from 'ngx-bootstrap/modal';
+
 import { ErrorModalComponent } from './modals/error-modal/error-modal.component';
 import { SuccessModalComponent } from './modals/success-modal/success-modal.component';
-import { DatepickerDirective } from './directives/datepicker.directive';
-
-
+import { AppendDirective } from './directives/append.directive';
 
 @NgModule({
-  declarations: [ErrorModalComponent, SuccessModalComponent, DatepickerDirective],
-  imports: [
-    CommonModule
+  declarations: [
+    ErrorModalComponent,
+    SuccessModalComponent,
+    AppendDirective
   ],
+  imports: [
+    CommonModule,
+    ModalModule.forRoot()
+  ], 
   exports: [
     ErrorModalComponent,
     SuccessModalComponent,
-    DatepickerDirective
+    AppendDirective
   ]
 })
 export class SharedModule { }
